v1.3

+++ final/frontend-react/src/component/Page/ViewBooking_User/ViewBookingUser.js
@@ -8,19 +8,20 @@ const cx = classNames.bind(styles);
 
 function ViewBookingUser() {
   const [meetingData, setMeetingData] = useState([]);
-  const pathBackEnd = ""
+
+  const pathBackEnd = "http://localhost:8081";
 
   useEffect(() => {
     const fetchMeetingData = async () => {
       try {
-        const response = await fetch(`${pathBackEnd}/getAllBookingByUserID/${localStorage.getItem("accountID")}`);
+        const response = await fetch(`http://localhost:8081/getAllBookingByUserID/${localStorage.getItem("accountID")}`);
         if (!response.ok) {
           throw new Error('Failed to fetch data');
         }
         const data = await response.json();
         setMeetingData(data.Data);
       } catch (error) {
-  
+
       }
     };
 
@@ -28,13 +29,17 @@ function ViewBookingUser() {
   }, []);
 
   const handleDeleteMeeting = (id) => {
-    fetch(`${pathBackEnd}/deleteMeeting/${id}`, {
-      method: 'DELETE',
-    })
+    const isConfirmed = window.confirm('Are you sure you want to delete?');
+    if (isConfirmed) {
+      fetch(`http://localhost:8081/deleteMeeting/${id}`, {
+        method: 'DELETE',
+      })
       .then(response => {
         if (response.ok) {
           // If the deletion is successful, update the local state
           setMeetingData(prevData => prevData.filter(meeting => meeting.id !== id));
+          alert('Deleted successfully ');
+
         } else {
           console.error(`Failed to delete meeting with ID ${id}`);
         }
@@ -42,7 +47,11 @@ function ViewBookingUser() {
       .catch(error => {
         console.error('Error deleting meeting:', error);
       });
+
     }
+  }
+
+
   return (
     <body>
       <section className={cx('hero')} >
@@ -52,35 +61,35 @@ function ViewBookingUser() {
           </h2>
           <div className={cx('text')}>We will respond to your request as soon as possible, please be patient </div>
           <div className={cx('BookingList')}>
-      <table>
-        <thead>
-          <tr>
-            <th className={cx('ID')}>ID</th>
-            <th className={cx('content')}>Content</th>
-            <th className={cx('timeRange')}>Time range</th>
-            <th className={cx('dateMeeting')}>Date Meeting</th>
-            <th className={cx('status')}>Status</th>
-            <th className={cx('delete')}>Actions</th>
-          </tr>
-        </thead>
-        <tbody>
-          {meetingData.map(meeting => (
-            <tr key={meeting.id}>
-              <td className={cx('ID')}>{meeting.id}</td>
-              <td className={cx('content')}>{meeting.content}</td>
-              <td className={cx('timeRange')}>{meeting.time_range}</td>
-              <td className={cx('dateMeeting')}>{meeting.date}</td>
-              <td className={cx('status')}>{meeting.status}</td>
-              <td className={cx('delete')}>
-                <button onClick={() => handleDeleteMeeting(meeting.id)}>
-                  Delete
-                </button>
-              </td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
+            <table>
+              <thead>
+                <tr>
+                  <th className={cx('ID')}>ID</th>
+                  <th className={cx('content')}>Content</th>
+                  <th className={cx('timeRange')}>Time range</th>
+                  <th className={cx('dateMeeting')}>Date Meeting</th>
+                  <th className={cx('status')}>Status</th>
+                  <th className={cx('delete')}>Actions</th>
+                </tr>
+              </thead>
+              <tbody>
+                {meetingData.map(meeting => (
+                  <tr key={meeting.id}>
+                    <td className={cx('ID')}>{meeting.id}</td>
+                    <td className={cx('content')}>{meeting.content}</td>
+                    <td className={cx('timeRange')}>{meeting.time_range}</td>
+                    <td className={cx('dateMeeting')}>{meeting.date}</td>
+                    <td className={cx('status')}>{meeting.status}</td>
+                    <td className={cx('delete')}>
+                      <button onClick={() => handleDeleteMeeting(meeting.id)}>
+                        Delete
+                      </button>
+                    </td>
+                  </tr>
+                ))}
+              </tbody>
+            </table>
+          </div>
         </div>
       </section>
     </body>
